Change data type for printHexDump function and update byte reading
The data type for the function `printHexDump` in `TerminalDisplay.java` is updated to use `int` array instead of `short` array. The byte reading logic in `Command.java` is also modified to read bytes into an `int` array before passing to `printHexDump`, improving the consistency and coherency of the reading and displaying process.

+++ src/TerminalDisplay.java
@@ -77,7 +77,7 @@ public class TerminalDisplay
         }
     }
 
-    public void printHexDump(short[] data, int baseAddress) {
+    public void printHexDump(int[] data, int baseAddress) {
         for (int i = 0; i < data.length; i += 8) {
             StringBuilder builder = new StringBuilder();
             builder.append(String.format("%04x  ", baseAddress + i));
